slp: remove from default images
slpd-lite[1] is a home grown application within OpenBMC that does not
appear to get much use. There are other more full function options like
Avahi or LLDP now. The slpd-lite daemon provides an unauthenticated
network interface which was the result of a fairly severe PSIRT last
year[2].

Given the above, this application should not be included by default when
building OpenBMC.

Tested:
- Confirmed slpd application and service no longer in a Romulus image

[1]: https://github.com/openbmc/slpd-lite
[2]: https://github.com/openbmc/slpd-lite/security/advisories/GHSA-wmgv-jffg-v3xr

Change-Id: Ifd67651cb00a411bf12f0330c4b4fab626953361
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>