meta-google: reconfigure interface upon networkd failure
Systemd 256 has issue with ndisc that local gateway is messed up, which
causes interface configuration to fail. The fix tries to reconfigure the
interface if that happens.

Also adds a timeout to ensure that we still proceed with doing RS/RA to
avoid regression.

Meanwhile the real underlay issue in systemd is not easy to root cause,
some upstream changes might be related but need time to investigate.

Tested: verified the script logic work, on a failed machine verified
that this can reconfigure the interface with no issue

Change-Id: Ie0acf183b055f2e293d8df8fcee325c083c6df1f
Signed-off-by: Yuxiao Zhang <yuxiaozhang@google.com>