meta-facebook: minerva: reload fan control when FCB plug-in
Add a fan reload script to automatically reload the fan
when it is unplugged and a new fan is plugged in.

For example: when we remove a fan in a sensor fail status and then plug
it back in, the fan speed will remain at 50% due to the CPLD default
setting. We reload the fan service because the PID controller will not
continuously send the same FAN PWM signal.

Change-Id: I6f05ff10bc654819bb2d33360a465ad619169c63
Signed-off-by: Yang Chen <yang.chen@quantatw.com>