meta-facebook: yosemite4: Remove TPM configs
These configs are already specified in tpm2.cfg from
the meta-aspeed layer which is enabled on `tpm2`
being present in the MACHINE_FEATURES which is true
for yosemite4 hence unnecessary to also provide in
machine specific config. In this particular case,
the machine config overrode the common config to install
TPM as a module which was wrong.

Tested: Built Yosemite4 machine and checked that
`/dev/tpm0` is created.

Change-Id: I36a5b9179b8b221d8ef8ce2f82217e21611e65ba
Signed-off-by: Amithash Prasad <amithash@meta.com>