meta-google: gbmc-bridge: Consolidate IP loading logic
We want to not only configure addresses but also set up the proper
RA advertisement parameters. This logic is the same between the bridge
static IP loading and the NCSI RA script. Consolidate into one function.

On top of this, we need to be able to load extra addresses at bootup.
The reload function will now consider lists of addresses in
`/run/gbmc-br-ips/*`

Change-Id: I3185fa5ade49d8d87de83ea3e3df60044a14ed78
Signed-off-by: William A. Kennington III <wak@google.com>