meta-google: gbmc-bridge: Find source with route metric
Without the metric, we can not determine if the route was intended to be
more strict based on our source policy. This fixes the logic to
guarantee that specific source routes for a gateway are preferred if an
interface has them.

Tested: Ran 4 powercycles and 3 reboots on a machine. Restarted
`gbmc-ip-monitor` and `systemd-networkd` a number of times to guarantee
stability in the result.

Change-Id: Ifb7b5d624d1e75070b3df00a57a074ba088a72d8
Signed-off-by: William A. Kennington III <wak@google.com>