meta-facebook: catalina: platform-sys-init: Remove unnecessary bind
When binding an I2C mux, any downstream devices defined in the device
tree (DTS) are automatically probed by the kernel once the mux is
initialized. Therefore, it is unnecessary to explicitly call
bind_i2c_device for components behind the I2C mux.

We observed that platform-sys-init fails to bind devices on systems with
2nd-source PDB boards due to an attempt to bind max31790 fan controller,
which is only present on the main-source board.

This change removes redundant I2C bindings for such devices to avoid
bind failures caused by differences between the main and 2nd-source
boards.

Change-Id: I757fbb1f407c6b6e36fc1a6d1ce6aed0dd6de0f1
Signed-off-by: Potin Lai <potin.lai@quantatw.com>