meta-evb-fvpbase: enable virtio-net for FVP
Enable virtio-net as an alternative NIC in FVP. FVP uses smsc_91c111
for NIC and its support is deprecated in uboot[1]. So, an
alternative NIC configuration is required for network access in uboot.

Testing:
Configured a TAP interface named 'tap1' for out-of-band access.
Sample setup on Ubuntu (requires sudo):

  sudo apt update
  sudo apt install qemu-kvm libvirt-daemon-system iproute2 bridge-utils
  sudo ip tuntap add dev tap1 mode tap user $(whoami)
  sudo ifconfig tap1 0.0.0.0 promisc up
  sudo brctl addif virbr0 tap1

Execute 'ip a' command in BMC shell and observe the NIC configuration.

Before:
  eth0 -> smsc_91c111

After:
  eth0 -> virtio-net
  eth1 -> smsc_91c111

References:
1. https://patchwork.ozlabs.org/project/uboot/patch/20220802113424.2988464-10-trini@konsulko.com

Change-Id: I445aecb6aefdb422a14fe131a64fc3de3b6865f9
Signed-off-by: Mohammed Javith Akthar M <mohammedjavitham@ami.com>