meta-phosphor:state-manager: use wants instead of requires
As noted in commit 7497224, when using a requires relationship with
multi-user.target, any restart of that service causes a restart of
multi-user.target, causing confusion with the BMC State and also
potentially starting some oneshot service that are not configured to
RemainAfterExit=yes. If the host or chassis services were to hit an
error or be restarted by a user then this situation occurs.

The requires relationship was initially used within
phosphor-state-manager to ensure a boot was not allowed if these
critical services were not started. Later the critical service monitor
was introduced to ensure the BMC state will got to Quiesced if they
fail (also, a system is not going to boot if the chassis or host
services are not running).

Tested:
- Booted up a system and killed phosphor-host-state-manager and verified
  we now only restart that one service

Change-Id: Ieebc05a747e738f195f58bdabcc12d0eed407d18
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>