meta-facebook: catalina: Handle error from standby-power-enable script
The platform-sys-init service failed because the standby-power-enable
script returned an error code (1), which prevented the PDB I2C Mux
driver from being bound correctly.

As a result, the BMC was unable to perform AC cycle operations via I2C
communication with HSC chips.

This commit adds error handling for the standby-power-enable script to
prevent the failure of the platform-sys-init service and ensure proper
initialization of the system.

Change-Id: I147acfb6c1d846b862d1711a7dd3bdce79bf345e
Signed-off-by: Potin Lai <potin.lai@quantatw.com>