meta-facebook: catalina: Remove exit on second FPGA ready check failure
Align with NVIDIA reference code of the second FPGA readiness check.
Instead of exiting on failure, Only logs a warning message if the second
FPGA is not ready, allowing the BMC to proceed with the remaining
initialization steps.

This ensures that a missing or delayed second FPGA does not block the
overall platform bring-up.

Change-Id: I715ea7c91ca9e135c415f7caa213446c28885123
Signed-off-by: Potin Lai <potin.lai@quantatw.com>