meta-google: networking: fix the primary route deletion
There is no deletion netlink message when changing the src address, so
we might miss deleting the primary path in some cases. This one deletes
the primary path when the src address is changed.

Tested: on a system tested that the primary path is cleared when the
primary interfaces are down.

Change-Id: I2ed4abd058dedb5ba1167b7d247f52d403fad432
Signed-off-by: Yuxiao Zhang <yuxiaozhang@google.com>