meta-facebook: harma: remove SetMtu
Observed recurring pldmTransport::sendMsg failures with response code
errno 113 - ECONNABORTED, likely due to a broken MCTP socket or a
dead endpoint. It seems to depend on busctrl to set the MTU value.
Maybe setting the MTU takes more time to initialize.

Setting the MTU was a workaround to avoid encountering a kernel issue.
Currently, this kernel issue has been fixed, so we can remove the
workaround.

'''
https://github.com/torvalds/linux/commit/70facbf978ac90c6da17a3de2a8dd111b06f1bac
'''

Change-Id: Id6ae37202ff3cd974f8fcda43353831da38cb714
Signed-off-by: Peter Yin <peter.yin@quantatw.com>