meta-facebook: catalina: fix bmc not ready
On the Catalina platform, the BMC fails to reach done state of the
`multi-user.target` because certain services are waiting for the
`/org/openbmc/control/power0` object path, which does not exist on this
platform, and it prevents BMC state become ready.

To resolve this issue, remove the dependencies on the following services
from systemd.

- obmc-power-start@.service
- obmc-power-stop@.service
- phosphor-reset-chassis-on@.service
- phosphor-reset-chassis-running@.service

Change-Id: Ie473bc48a83f663009a55bb5989c3a3b8497092f
Signed-off-by: Potin Lai <potin.lai@quantatw.com>