meta-facebook: yosemite4: Removed the fixed MTU 64 setting from the MCTP config
# Description
Fixed MCTP configuration: Changed MTU (Maximum Transmission Unit) from a fixed 68 bytes
to the default setting. According to the MCTP specification, if the MTU is not explicitly
set, it defaults to the MTU of the network card.

# Motivation
According to the DFT specification, the WF CXL update time meets the requirement of completing
within five minutes. This modification reduces the CXL firmware update duration to under five minutes.

Change-Id: Ice3e80618ec8a821871288c4814dd671008127f0
Signed-off-by: Carter Chen <carter.chen.wiwynn@gmail.com>