meta-facebook: yosemite4: fix bmc always not ready
Related dependencies for these power services were already removed from
CHASSIS_DEFAULT_TARGETS, so they should not be needed on the Yosemite4
platform.
However, the services were still present and waiting for D-Bus object
paths (such as /org/openbmc/control/power0) that do not exist on this
platform, causing the BMC not to reach the ready state.

This change removes those unused services:
- obmc-power-start@.service
- obmc-power-stop@.service
- phosphor-reset-chassis-on@.service
- phosphor-reset-chassis-running@.service

Change-Id: I14cb962d19d1e4a25c956e2433d30ab438b4104b
Signed-off-by: Eric Yang <eric.yang.wiwynn@gmail.com>