meta-facebook: ventura: Enable the small leak LED table behavior
According to the LED table defined in the specification, the amber
light must turn on in the event of a small leak. Therefore, enable
this function following the LED table definition.

We also added a check to ensure both the Big and Small LeakDetector
statuses are Normal before turning off the port’s Leak LED.

Change-Id: I6df967534f29313adc68537ec88953808b7eadc3
Signed-off-by: Jason Hsu <Jason-Hsu@quantatw.com>