meta-facebook: darwin: Add empty leds configuration
Unless we have a yaml / json description file for the LEDs, phosphor-led-manager's service "xyz.openbmc_project.LED.GroupManager.service" will not finish running, which messes up with the internal states of phosphor-state-manager.

~# systemctl list-jobs
JOB  UNIT                                                TYPE  STATE
380  obmc-led-group-start@bmc_booted.service             start waiting
385  phosphor-discover-system-state@0.service            start waiting
360  mapper-wait@-xyz-openbmc_project-led-groups.service start running
221  multi-user.target                                   start waiting
359  obmc-fru-fault-monitor.service                      start waiting
1645 xyz.openbmc_project.LED.GroupManager.service        start running

In Darwin, the LEDs are controlled by the host, not by the BMC, so leave the config file empty.


Change-Id: I43e6ad89c3a3180f91bb42e0c7d7ccf6d4d7d82b
Signed-off-by: Robert Ercean Dima <robert3@meta.com>