meta-facebook: clemente: Modify tray leak event handling behavior.
In the previous architecture, when a tray leak event occurs,
the PDB is responsible for controlling the HPM to shut down the host,
and the BMC only needs to send an event message.

However, in the new hardware architecture,
the interposer cannot directly control the HPM to shut down the host,
and it must rely on the BMC to notify the SCM CPLD to control the HPM
for host shutdown.

Tested: Verified host shutdown after receiving the tray leak event.

Change-Id: Ie4234b9eccdc05997100f839b5ae84cdc5068578
Signed-off-by: Kimi Chen <kimi.zy.chen@fii-foxconn.com>