meta-ibm: romulus: Disable DRM to reduce v6.12 binary size
Shrinking vmlinux by disabling CONFIG_DRM allows the resulting FIT image
to meet the size constraint imposed by its flash partition.

CONFIG_DRM and CONFIG_DRM_ASPEED_GFX tend to be used for fbterm and
rendering the host console managed by uart-render-controller. For some
IBM platforms this is necessary, but not so much for Romulus. Disable it
to reduce the FIT image size so that it is not too large for its flash
partition.

For comparision, size(1) reports:

    dev-6.6:
       text    data     bss     dec     hex filename
    7134648 2425590  144288 9704526  94144e vmlinux

    dev-6.12, CONFIG_DRM=y:
       text    data     bss     dec     hex filename
    7504840 2533774  160576 10199190 9ba096 vmlinux

    dev-6.12, CONFIG_DRM=n:
       text    data     bss     dec     hex filename
    7117484 2399170  158304 9674958  93a0ce vmlinux

Suggested-by: Joel Stanley <joel@jms.id.au>
Change-Id: I87b10a538c841a2a74861ff49cd497c8be173ccf
Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>