meta-facebook: catalina: add ignoreFailIfHostOff setting for FSC config
Add the ignoreFailIfHostOff setting for sensors that only provide valid
readings when the host is powered on. This allows the FSC service to
treat unavailable readings as acceptable when the host is off, while
still triggering failsafe mode if the readings are unavailable when the
host is on.

This change relies on the following commit:
[1] https://gerrit.openbmc.org/c/openbmc/phosphor-pid-control/+/83200

Change-Id: I0967f1bdc12dcdab99c4b37d6c56588ad01d4a98
Signed-off-by: Potin Lai <potin.lai@quantatw.com>