meta-google: gbmc-bridge: Support multiple DHCPv6 addresses
We will now be sending all of the addresses for a node via DHCP, parse
all of them and configure them on the gbmcbr.

Tested: Run though an installation with manual breakpoints after IP
setting to verify that the BMC has the correct IP information
provisioned and is accessible via both paths.

Change-Id: I90b36710bd7478273c23324f86f1676311861143
Signed-off-by: William A. Kennington III <wak@google.com>