meta-facebook: harma: set power off before ac cycle
According to JEDEC (JESD301-1A), the PMIC requires a minimum
Vin_Bulk voltage of 5 V and a duration of at least 200 ms after
DIMM_CAMP assertion to guarantee successful writes to non-volatile
memory. However, the motherboard currently discharges too quickly.
To address this, the gpio signal is set to trigger the DC power
drop, giving the CPLD enough time to inform the DIMM and CPU before
the AC cycle, ensuring that the write operation completes reliably.

Change-Id: Ic79740b8fcde90d78b0ecd4678f48922a1cf5b65
Signed-off-by: Peter Yin <peter.yin@quantatw.com>