meta-facebook: ventura: auto-update RS485 udev rules
Since the number of USB-to-RS485 chips differs between RMCv1 and RMCv1+,
placing both rs485 port mapping rules in the same udev rules file may
lead to conflicts. The same rule could result in different port mappings
depending on the RMC revision.

To address this issue, the number of ttyUSB* devices is counted to
identify the RMC hardware revision, and the correct udev rules are then
applied automatically.

Tested:
We have verified that the rules work on the RMCv1 and RMCv1+ machines.

Change-Id: Ic0aa011871a3cbf19884e8b45172c72466220745
Signed-off-by: P.K. Lee <p.k.lee@quantatw.com>