meta-facebook: yosemite4: Add host powerreset failure service
Description:
When DC/AC is off, executed host powerreset results in status mismatch
on D-Bus. The root cause is that the host-reboot service continues to run,
causing the power state to appear inconsistent.

Motivation:
To ensure power reset is executed while the DC is ON. Address
power state mismatches after power reset when DC/AC is off.

Tested:
- Check power state after power control - pass
- Execute power reset when DC/AC is off, then check whether the
power state is consistent.

Change-Id: I6f7bc0698b80cf29ed383c8f8f05d3af9043f2ff
Signed-off-by: Zoey YJ Chung <zoey.yj.chung.wiwynn@gmail.com>