meta-facebook: yosemite5: add setup OS status services
Add host ready assert and deassert services to the GPIO monitor to
update the OS status property on D-Bus.
Add a multi GPIO initialization service to set the OS status property
during BMC boot.

These services set the OS status property on D-Bus to indicate when
the host has reached a ready or not-ready state.

Change-Id: Ifb436da40530e968ebd6f6c6da7558bb098f3f4f
Signed-off-by: Kevin Tung <kevin.tung.openbmc@gmail.com>