phosphor-pid-control: Add `handle-missing-object-paths` to PACKAGECONFIG
This commit adds the feature `handle-missing-object-paths` to the
PACKAGECONFIG list of phosphor-pid-control which can enable a meson
option of the same name. This option is supported by the commit [1] and
is false by default. The feature is not enabled by this commit. If
enabled, swampd will add more behaviors to the current handling of
`MissingIsAcceptable` config, which is: When the input config is from
D-Bus and a dbuspassive input sensor is missing from D-Bus on init or
rescan, this sensor will be built as a failed sensor and will send the
zone to failsafe mode. If the sensor comes back to D-Bus, swampd will
be informed through the existing match signals registered. Behaviors
when the config is parsed from the legacy JSON file are not affected.

[1]: https://gerrit.openbmc.org/c/openbmc/phosphor-pid-control/+/76062

Change-Id: I6cf5ba1e81722888987655a8063ff30d68b95dc3
Signed-off-by: Chau Ly <chaul@amperecomputing.com>