meta-facebook: catalina: align RequestHostTransition with CurrentHostState
Ensure the `RequestHostTransition` property is updated to match the
`CurrentHostState` property. The BMC uses `RequestHostTransition` to
trigger obmc-host-start/shutdown targets, so these properties are
expected to stay aligned.

A service may invoke the obmc-host-shutdown target when a fault is
triggered. For example, phosphor-fan-monitor does so when a fan fault
occurs. This can cause a mismatch between `RequestHostTransition` and
`CurrentHostState`, which may lead the BMC to misjudge the host power
state when it is powered on via the power button.

This commit adds a check to ensure the `RequestHostTransition` property
is always aligned with `CurrentHostState`.

Change-Id: I9a939a1d234f29432f17bb6c2cb7eec0fad66b23
Signed-off-by: Potin Lai <potin.lai@quantatw.com>