meta-google: obmc-phosphor-initfs: whitelist: Add random-seed
We want to keep the random-seed between boots, as we use it not only for
seeding RNG but we also for updating the initial system time. This fixes
our initial time prior to having NTP sync.

Tested: Ran on a BMC to verify that it kept the file across reboot, this
allows the time to better be retained between boots.

Change-Id: Ifa78bf41baac9a8165715cbe4b0ae8f6a9db1649
Signed-off-by: William A. Kennington III <wak@google.com>