meta-facebook: ventura: Add RMC notification of LSC port ID via I2C
The RMC sends the currently connected LSC port ID to each compute tray
and switch tray via the LSC I2C bus when the RMC boots or when a tray
powers on. From the RMC’s perspective, both a tray power-on event and
an LSC port change trigger the same condition, as the RJ45 power-good
pin transitions from off to on.

This allows each tray to verify whether it is connected to the correct
LSC port based on the received ID.

Change-Id: Iedcdfa658a1bc1053df453381d5fb6c0d31b3b0f
Signed-off-by: Fred Zhang <fred.zhang@quantatw.com>