ipmitool: fix malloc errors during fru edit
Ipmitool fru edit command is failing with below malloc errors
and core-dumps are getting generated. It is happening while editing
the fru for certain sizes. Errors are like below logs:

malloc_consolidate(): invalid chunk size
Aborted (core dumped)

Existing patch has wrongly implemented the moving of fru data in
MultiRecord area while editing the fru area. Modified the existing
patch to fix this issue. Fixed the record length of MultiRecord
area as per the ipmi-platform-mgt-fru-info-storage specification.
The record length is not multiple of 8 bytes, it is the exact number
of bytes of data in the record. Added 5 to the length as header size
is of 5 bytes.

Tested:
ipmitool fru edit 0 field b 1 "ABC DEF HI212345678"

Change-Id: I0e1709da4cdd5ff0c3743708f03eedad64bb3bfd
Signed-off-by: Jitendra Tripathy <jitendra.kumarx.tripathy@intel.com>