meta-google: support all kinds of machine assignment
This change enables less strictness on matching regex to ignore machine name and does not have to match. We do not need to do that anymore as our systems assign packages properly and we need flexibility to move between different os-release machine names. Further, added exclusion regex ability in function to enable users to exclude certain matches if necessary.

Change-Id: I0f03cbf57902bb167477a9ab746b75653c8dad25
Signed-off-by: Mo Elbadry <elbadrym@google.com>