meta-facebook: ventura: Add checked for the DUID-LL configuration
The root cause is that during BMC boot, the Entity Manager has not yet
written the MAC address to the Inventory when the Network service start.
This results in the error “Exception occurred during getting of MAC
address from Inventory” and triggers a restart of the Network service.

However, the file 00-bmc-eth0.network is already reconfigured with the
DUID-LL settings during the first Network service start. When the Network
service restart, the file is overwritten and ends up without the DUID-LL
settings, which ultimately leads to the DHCPv6 failure.

Therefore, we added an extended check function for the DUID-LL settings
to ensure that after the Network service restart, the 00-bmc-eth0.network
file is also checked for the presence of the DUID-LL configuration.

Change-Id: I4c96ee6bd5abed9d40c462f53b512893be338826
Signed-off-by: Jason Hsu <Jason-Hsu@quantatw.com>