phosphor-health-monitor: Adjust Storage_RW monitoring path for phosphor-mmc
The default monitoring path for Storage_RW is /run/initramfs/rw. However,
phosphor-mmc mounts the RW filesystem on the /var path, so the path
in bmc_health_config.json needs to be updated.

The changes added here to replace the default monitoring path that trigger
by the df-phosphor-mmc feature distro. And also allow other recipes to
override the JSON file if needed.

Change-Id: I2682de7a9e6c96c8c0fb01db3fe43a82f5d576d7
Signed-off-by: John Chung <john.chung@arm.com>