meta-ibm: introduce ibm-enterprise machine override
As new IBM machines are defined it has become obvious that the use of
p10bmc has started to be used for a lot more then just p10 bmc machines.
Really, it's the first machine that defined our IBM enterprise
functions. Convert most of the "p10bmc" machine functions over to a new
ibm-enterprise machine override (similar to the ibm-ac-server override).

This machine override will be used for all of the shared IBM enterprise
functions across all IBM systems.

This commit left p10bmc for the hwmon/sensor area, IPMI,
and a few others which were specific to the p10 configurations.

Tested:
- Confirmed p10bmc image still builds

Change-Id: I78fb90aa8d96672984c672eae44242c877144cda
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>