meta-facebook: clemente: add OSFP transceiver temperature sensors to PID config
This change adds monitoring support for the OSFP transceiver temperature
sensor in the phosphor-pid-control configuration.

To support the differences between NSF (Non-Scheduled Fabric) and DSF (Decoupled
Scheduled Fabric) systems, the OSFP temperature sensor is marked as
missingIsAcceptable. OSFP transceiver sensors are not configured on NSF platforms;
without this setting, the system will enter fail-safe mode due to lack of input.

Change-Id: Ica579f372500be03094f28c5fa4dfda52d8fd93e
Signed-off-by: Leo Wang <leo.jt.wang@fii-foxconn.com>