yosemite4: Adjust CPU power status check in check-interrupt
Description:
Adjust  CPU power status check so that P0 power good is checked
only when power fault occurs, and the D-Bus state is updated to
the correct status. Retain the original conditional logic to ensure
that any remaining state inconsistencies are still recorded in the
journal logs.

Motivation:
When execute ACPI reset, the CPU may have an low pulse on P0 power
good. This trigger BMC interrupts (INT_SMB_BMC_SLOT1_4_N or
INT_SMB_BMC_SLOT5_8_N).In check-interrupt, it observes that the
P0 power good status is 0 and inconsistent with the D-Bus state,
which may cause the power state to unexpectedly transition to off
and result in a DC off during the ACPI reset.

Tested:
- Check power state after acpi reset - pass
- Stress ACPI reset and ensure that DC off does not occur.

Change-Id: Ic0264eec945ec3cace81e885075e3ea09743775c
Signed-off-by: Zoey YJ Chung <zoey.yj.chung.wiwynn@gmail.com>