meta-phosphor: phosphor-static-norootfs-init: commonize flash-bank detection
When there are two SPI-NOR banks, it is helpful to leave a file in
`/run` to identify which bank has been booted from.  Most platforms
already have a copy of a snippet of code that reads from the watchdog
to determine which bank has been booted from.  Make this code common.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: Id516fd28234f59ed7718ca3cc7a05ce0a21035e5