meta-facebook: harma: sgpio 129 to default value
Before issuing the AC cycle command, the BMC may trigger a watchdog
timeout and reset unexpectedly.

In this case, the SGPIO output value read from the CPLD does not
return to the expected default state after reboot, causing incorrect
SGPIO status to be reported.

As a workaround, force GPIO129 to low during GPIO initialization to
ensure SGPIO starts from a known default state after a watchdog-
triggered reset.

This workaround is required when the reset occurs before the AC
cycle command is executed.

Change-Id: Ib60a4ee0c68cfaa6c393f1b06c62e4264427956d
Signed-off-by: Peter Yin <peter.yin@quantatw.com>